Prefactor: state the type-compatibility relation once, outside the function-call rule
The predicate that decides whether a value's resolved type is acceptable
where another type is required (exact match, the ANY_* categories that
standard-library signatures use, untyped-literal inference, lossless and
flag-gated widening) lived inside rule_function_call_type_check as a set
of private functions, although the rule already applied it to three
different positions: an argument against its parameter, a function
result against its assignment target, and an assignment value against
its variable. A rule that needs to ask the same question of an operator's
operand had no way to reach it without copying it.

The predicate and its unit tests now live in their own module,
type_compat, crate-visible, and the function-call rule imports it. The
helper that says which type names the predicate can judge is renamed
from is_checkable_assignment_type to is_checkable_type, since nothing
about it is specific to assignment. Behaviour is unchanged; every test
passes as it was, and the function-call rule module drops from 1801
lines to 1459.
